Decouple InfiniteScrollTable from track selection.
-Decouple InfiniteScrollTable from track selection.
