+++ src/view/src/widgets/rocprofvis_infinite_scroll_table.h
@@ -4,7 +4,6 @@
 
 #include "imgui.h"
 #include "rocprofvis_data_provider.h"
-#include "rocprofvis_events.h"
 #include "widgets/rocprofvis_widget.h"
 #include <string>
 #include <vector>
@@ -14,49 +13,31 @@ namespace RocProfVis
 namespace View
 {
 
+constexpr const char* ROWCONTEXTMENU_POPUP_NAME = "RowContextMenu";
+
 class SettingsManager;
-class TimelineSelection;
+class TableDataEvent;
 
 class InfiniteScrollTable : public RocWidget
 {
 public:
-    InfiniteScrollTable(DataProvider&                      dp,
-                        std::shared_ptr<TimelineSelection> timeline_selection,
-                        TableType table_type = TableType::kEventTable);
-
+    InfiniteScrollTable(DataProvider& dp, TableType table_type = TableType::kEventTable);
     void Update() override;
     void Render() override;
     void SetTableType(TableType type) { m_table_type = type; }
 
-    void HandleTrackSelectionChanged();
     void HandleNewTableData(std::shared_ptr<TableDataEvent> e);
 
-private:
-    // Important columns in the table
-    enum ImportantColumns
-    {
-        kId,
-        kName,
-        kTimeStartNs,
-        kTimeEndNs,
-        kDurationNs,
-        kTrackId,
-        kStreamId,
-        kNumImportantColumns
-    };
-
+protected:
     struct FilterOptions
     {
         int  column_index;
         char group_columns[256];
         char filter[256];
     };
-    void RenderContextMenu() const;
-    void RenderLoadingIndicator() const;
-    bool XButton(const char* id) const;
 
-    uint64_t GetTrackIdHelper(
-        const std::vector<std::vector<std::string>>& table_data) const;
+    virtual void FetchData(const TableRequestParams& params) const = 0;
+    virtual void RenderContextMenu() const;
 
     std::vector<std::string> m_column_names;
     std::vector<const char*> m_column_names_ptr;
@@ -66,28 +47,27 @@ private:
     TableType m_table_type;  // Type of table (e.g., EventTable, SampleTable)
     rocprofvis_controller_table_type_t m_req_table_type;
 
-    DataProvider&                      m_data_provider;
-    SettingsManager&                   m_settings;
-    std::shared_ptr<TimelineSelection> m_timeline_selection;
+    DataProvider&    m_data_provider;
+    SettingsManager& m_settings;
 
     int m_fetch_chunk_size;
+
+    bool m_data_changed;
+
+    // Track the selected row for context menu actions
+    int m_selected_row = -1;
+
+private:
+    void RenderLoadingIndicator() const;
+    bool XButton(const char* id) const;
+
     int m_fetch_pad_items;
     int m_fetch_threshold_items;
 
     // Internal state flags below
     bool     m_skip_data_fetch;
     uint64_t m_last_total_row_count;
-    bool     m_data_changed;
-    bool     m_defer_track_selection_changed;
-
-    // Track the selected row for context menu actions
-    int    m_selected_row = -1;
-    ImVec2 m_last_table_size;
-
-    // Keep track of currently selected tracks for this table type
-    std::vector<uint64_t> m_selected_tracks;
-
-    std::vector<size_t> m_important_column_idxs;
+    ImVec2   m_last_table_size;
 };
 
 }  // namespace View
